Modify output method of loggingT to handle Logger/InfoLogger correctly
This commit modifies output method of loggingT to handle l.logr
or argument log correctly, if set. And add test function of
output method of loggingT.

In this commit l.logr is checked first, because if loggingT.logr,
which is only receiver value of output method currently, is set,
it should be used unconditionally as comment of logr field of
loggingT says.

Till this commit applies, output method is only used with global
variable logging as receiver and logging.logr/Verbose.logr as
argument log.
To maintain backward compatibility, in this commit, output method
only checks argument s is errorLog or not when l.logr is not nil.