add structured logging helper
When passing an object with complex content as a key in a key/value list, good
formatting of that object will be different for text and structured
output. Custom formatters can be provided by implementing fmt.Stringer and
logr.Marshaler, but that implies defining custom types outside of the function
which does the logging.

The new klog.As helper struct provideds the necessary boilerplate code and allows
users to keep all code inside the same function.