extract values from context on demand
Having to add a modified logger to a context when adding some values
to the context which are meant to be logged (like trace ID) has two
drawbacks:
- the instrumentation code which adds the values to the context must
  be aware of logging
- modifying the logger incurs a cost, whether some actual log entry
  then gets emitted or not.

A better approach is to add the values only to the context, then during logging
extract them. The solution in this commit does that entirely in klog: it wraps
the LogSink from the context or the default in a LogSink which remembers the
context and the desired keys, then adds them to the parameters.

Adding to the parameters was chosen instead of using WithValues because it was
a bit faster.

With this approach, doing calls with no logging because of the verbosity
threshold (100PercentDisabled) and only emitting a log entry for 1% of the
iterations (1Percent) becomes considerably cheaper and faster. The case where
log entries really get emitted becomes slower, but that is acceptable because
actually doing logging is already slow and not necessarily the case that we
must optimize for.

name                                   time/op
NewContext1Percent-36                  67.8µs ± 4%
NewContext100PercentDisabled-36        86.4µs ± 3%
NewContext100Percent-36                90.7µs ± 2%
NewContext1PercentValues-36            27.0µs ± 2%
NewContext100PercentDisabledValues-36  64.3µs ± 2%
NewContext100PercentValues-36           122µs ± 2%

name                                   alloc/op
NewContext1Percent-36                  23.3kB ± 0%
NewContext100PercentDisabled-36        32.8kB ± 0%
NewContext100Percent-36                32.8kB ± 0%
NewContext1PercentValues-36            10.0kB ± 0%
NewContext100PercentDisabledValues-36  25.6kB ± 0%
NewContext100PercentValues-36          48.0kB ± 0%

name                                   allocs/op
NewContext1Percent-36                     501 ± 0%
NewContext100PercentDisabled-36           600 ± 0%
NewContext100Percent-36                   600 ± 0%
NewContext1PercentValues-36               205 ± 0%
NewContext100PercentDisabledValues-36     400 ± 0%
NewContext100PercentValues-36             700 ± 0%