support "structured error"
A "structured error" is an error type that implements both Error and
MarshalLog. It gets logged like a normal error with "err=<Error()>", but then
another "errDetails=<MarshalLog()>" gets added to log additional information
that may be stored in the error.

The result of MarshalLog is logged like any other value. Beware that rendering
of structs, in particular multi-line strings in structs, is not very readable
because it all gets handled by `fmt.Sprintf`. To get nicer output of multiple
values, a logr.KeysAndValues can be returned. Those values than are formatted
one-by-one by klog, which means that multi-line strings are readable.