support "structured error"
A "structured error" is an error type that implements both Error and
ErrorDetails. It gets logged like a normal error with "err=<Error()>", but then
another "errDetails=<ErrorDetails()>" gets added to log additional information
that may be stored in the error.

The result of ErrorDetails is logged like any other value. Beware that rendering
of structs, in particular multi-line strings in structs, is not very readable
because it all gets handled by json.Marshal. This also means that unexported
fields are not logged.

To get nicer output of multiple values, a PseudoStruct can be returned. Those
values then are formatted one-by-one by klog, which means that multi-line
strings are readable.