klogr: avoid calling Info/Error/Enabled through auto-generated code
These functions get called for a pointer because that is what we use to
implement the LogSink interface. The variant of these functions which take a
value instead of a pointer was never used, so providing directly what is needed
instead of relying on the compiler to generate that is more efficient (less
code, avoids one extra callstack entry at runtime).

The With* functions must continue to take a value because they need to modify
that copy.