programmatically set oneoutput flag
This enhancement enables an application to programatically set the oneOutput flag
without needing to expose flags on the command line. It aligns with the existing
functions LogToStderr and SetOutputBySeverity, such that one can direct messages
to different output destinations and control whether those messages are only
written to their native severity level.