ktesting: fix vmodule support
Stack unwinding to determine the caller must consider the number of call levels
to skip over in logr itself. This only affected the `Enabled` check and thus
usage of the vmodule parameter, the actual caller then gets printed by
testing.T.

This also leads to a small inconsistency: the enabled check is unaware of the
testing.T.Helper calls and can only check the direct caller, without skipping
helpers, while the actual log output then skips the helper. There's no good fix
for this, testing.T would need an API for delegating the stack unwinding to the
caller (would also be useful to log a PC supplied via slog...).